refactor GFrameHoldeContext as a member of EntryFrame
