+++ src/musx/dom/Details.h
@@ -707,90 +707,6 @@ public:
 };
 
 /**
- * @class GFrameHoldContext
- * @brief A context wrapper for @ref GFrameHold associated with a specific part and location.
- *
- * This class retrieves the appropriate @ref GFrameHold from a Document using part, instrument, and measure IDs,
- * and enables part-aware operations like iterating over EntryFrame objects.
- */
-class GFrameHoldContext {
-public:
-    /**
-     * @brief Constructs a context-aware @ref GFrameHold wrapper.
-     * 
-     * @param document Weak pointer to the owning Document.
-     * @param partId The requested part ID.
-     * @param inst The instrument ID for.
-     * @param meas The measure ID for.
-     */
-    GFrameHoldContext(const DocumentPtr& document, Cmper partId, Cmper inst, Cmper meas);
-
-    /**
-     * @brief Returns the requested part ID associated with this context.
-     * 
-     * @return The requested part ID.
-     */
-    Cmper getRequestedPartId() const { return m_requestedPartId; }
-
-    /**
-     * @brief Provides const pointer-style access to the underlying @ref GFrameHold.
-     * 
-     * @return A const pointer to @ref GFrameHold.
-     */
-    const GFrameHold* operator->() const { return m_hold.get(); }
-
-    /**
-     * @brief Returns true if the internal @ref GFrameHold is valid.
-     * 
-     * @return True if the @ref GFrameHold was successfully retrieved; false otherwise.
-     */
-    explicit operator bool() const { return static_cast<bool>(m_hold); }
-
-    /// @brief Returns the clef index in effect for at the specified @ref Edu position.
-    /// This function does not take into account transposing clefs. Those are addressed in #createEntryFrame.
-    /// @param position The Edu position of the clef *in staff-level Edus*. (The staff-level matters for Independent Key Signature staves.)
-    ClefIndex calcClefIndexAt(Edu position) const;
-
-    /// @brief Returns the clef index in effect for at the specified @ref util::Fraction position (as a fraction of whole notes).
-    /// This function does not take into account transposing clefs. Those are addressed in #createEntryFrame.
-    /// @param position The *staff-level* position of the clef. (The staff-level matters for Independent Key Signature staves.)
-    ClefIndex calcClefIndexAt(util::Fraction position) const
-    { return calcClefIndexAt(position.calcEduDuration()); }
-
-    /** @brief Returns the @ref EntryFrame for all entries in the given layer.
-     *
-     * @param layerIndex The layer index (0..3) to iterate.
-     * @param forWrittenPitch If true, the key and clef for each entry are calculated for written pitch rather than concert pitch.
-     * @return EntryFrame for layer or nullptr if none.
-     */
-    std::shared_ptr<const EntryFrame> createEntryFrame(LayerIndex layerIndex, bool forWrittenPitch = false) const;
-    
-    /**
-     * @brief iterates the entries for the specified layer in this @ref GFrameHold from left to right
-     * @param layerIndex The layer index (0..3) to iterate.
-     * @param iterator The callback function for each iteration.
-     * @return true if higher-level iteration should continue. false if it should halt.
-     * @throws std::invalid_argument if the layer index is out of range
-     */
-    bool iterateEntries(LayerIndex layerIndex, std::function<bool(const EntryInfoPtr&)> iterator);
-
-    /**
-     * @brief iterates the entries for this @ref GFrameHold from left to right for each layer in order
-     * @param iterator The callback function for each iteration.
-     * @return true if higher-level iteration should continue. false if it should halt.
-     */
-    bool iterateEntries(std::function<bool(const EntryInfoPtr&)> iterator);
-
-    /// @brief Calculates the number of voices used by the GFrameHold instance.
-    /// @return A list of each layer that contains entries and whether that layer uses voice2.
-    std::map<LayerIndex, bool> calcVoices() const;
-
-private:
-    std::shared_ptr<GFrameHold> m_hold;      ///< The resolved GFrameHold object, or null if not found.
-    Cmper m_requestedPartId;                 ///< The requested part context.
-};
-
-/**
  * @class IndependentStaffDetails
  * @brief Represents independent time and key signature overrides for a staff.
  *
